reddit_ingestion_cf/requirements.txt
praw
google-cloud-pubsub
google-cloud-storage
python-dotenv
pytest